+++ Gravitas.Platform.Web/ViewModel/OpRoutine/LoadPointGuide/02_BindLoadPoint.cs
@@ -6,7 +6,7 @@ namespace Gravitas.Platform.Web.ViewModel
     {
         public class BindDestPointVm
         {
-            public long NodeId { get; set; }
+            public int NodeId { get; set; }
 
             [DisplayName("Картка")]
             public string Card { get; set; }
@@ -27,7 +27,7 @@ namespace Gravitas.Platform.Web.ViewModel
             public bool IsThirdPartyCarrier { get; set; }
 
             [DisplayName("Точка призначення")]
-            public long DestNodeId { get; set; }
+            public int DestNodeId { get; set; }
 
             public string DestNodeName { get; set; }
